refactor integration test namespace
